Overhauled Player and Enemy classes (and the enemy factory), turning them into interfaces and adding ActivePlayer and CurrentEnemy to replace the original classes. Started implementing the InBattleDataAccessObject. The Character interface can now probably be deleted, but it has been left intact for now, just in case.
