+++ src/data_access/InBattleDataAccessObject.java
@@ -1,0 +1,11 @@
+package data_access;
+
+import entity.Enemy;
+import entity.Player;
+
+public class InBattleDataAccessObject {
+
+    private Player player;
+    private Enemy enemy;
+
+}
